RelayServer 对齐进程分发与运行时签名
- 对齐 XRelayServer 与 XLoginServer 的公共运行时接口，统一 XServer::Clear、TXDBSocket::FindUser、SetMyInfo、SendUpdateServerInfo 等核心签名
- 调整 RelayServer 进程分发模型，统一 Server/User/Friend/Party/Force/League/ModeMaze 处理器为无参构造与注册后初始化，并补齐访问级别和参数类型
- 新增 MonitorProcess，补上 main=0xF7 的监控请求分发与 ServerInfo 汇总返回链路
- 补齐 CServer 的同步加载状态定义，使用 E_SERVER_SYNC_LOAD 枚举替代硬编码，并修正 RelayControlSocket/ServerProcess 的相关接线
- 完善 UserObject 与 CCommunity 的缺失方法和签名，包括 SetServer、Logout、Levelup、UpdateProfilePhoto、SendUpdateCommunity、ChangeFriendName、LoadBlock、AddBlockList、UpdateCharCommunity 等对象层行为
- 修正 PartyManager 缺失的 GetParty(UXActorID) 重载，并同步调整 Force、Party、Friend、League 各 Process、Manager、Matching 的调用方式以匹配 IDA
- 调整 League 相关按值传参与申请列表维护逻辑，补上 ATL::CTime/CTimeSpan 支撑、申请过期判断以及 m_mpLeagueApplicant 字段对齐
- 同步修正 LoginServer/GameDBSocket 的会话查找、角色切服/换图回包与 Relay 控制链路接口，保持 RelayServer 与 LoginServer 构建通过

+++ F/_PROGRAM_HG/Source/Soulworker/GameServer/XCore/XServer/TXDBSocket.h
@@ -1955,7 +1955,7 @@ public:
         int xSessionID = 0;
         xPacket.XParse >> xSessionID;
 
-        TUser* pUser = FindUser(static_cast<unsigned int>(xSessionID));
+        TUser* pUser = FindUser(xSessionID);  // 对齐 IDA: int 参数
         if (pUser && !pUser->GetBlockType()) {
             return DBParse(pUser, xPacket);
         }
@@ -1971,7 +1971,8 @@ public:
     }
 
 protected:
-    virtual TUser* FindUser(unsigned int xSessionID) {
+    // 对齐 IDA: 参数类型 H = int
+    virtual TUser* FindUser(int xSessionID) {
         (void)xSessionID;
         return nullptr;
     }
